List edited
The class List was edited in order to add a method that will add the list of adjacents based on the ID we need to look for.
Change the private attributes of List to public in order to use the pointers.
The method printList was edited in order to print the original list and the adjacents of each country.