List updated.
The list was updated in order to change the pointer to a nother list for a pointer to a vector of pointers, these pointers will be pointers to the adjacent countries.
That way if something chages in a ponter, it will change for the actual country and all the pointers to it.